week-4/Day-4-Thursday/labs/lab-10-techmart-etl/student_distribution/requirements.txt
# Core ETL Pipeline Dependencies
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
pymongo>=4.0.0
pyodbc>=4.0.30

# Data Quality and Validation
pydantic>=1.10.0
jsonschema>=4.0.0
great-expectations>=0.15.0

# Testing Framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0
pytest-asyncio>=0.21.0

# Configuration and Environment
python-dotenv>=0.19.0
pyyaml>=6.0.0
click>=8.0.0

# Database Drivers
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
pymssql>=2.2.0

# Web and API Support
requests>=2.28.0

# Time and Date Handling
python-dateutil>=2.8.0

# Monitoring and Logging
structlog>=22.0.0
prometheus-client>=0.14.0
psutil>=5.9.0

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0
lxml>=4.9.0

# Development Tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
